backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg[binary]==3.1.19
python-dotenv==1.0.1
pytest==8.3.2
pytest-django==4.8.0
django-cors-headers==4.3.1
